pi/mining/requirements.txt
paho-mqtt
python-json-logger